Upgrade to Muzei API 3.4
Muzei API 3.4 significantly improves the performance of `addArtwork` when adding multiple artwork (such as what `PhotosWorker` does, which got slightly updated to take into account the immutability of `Artwork` now).

Muzei API 3.4 depends on Kotlin 1.3.72 and a `compileSdkVersion` of 30, which enforced an upgrade to Android Gradle Plugin 4.0.1 (and with that, Gradle 6.5) and required a nullability fix in `SettingsActivity`.

The upgrade also affected how actions are implemented in `PhotosArtProvider` by deprecating the `getCommands()` and `onCommand()` methods (which are kept now only for backward compatibility with Muzei 3.3 users) and replacing them with the new `getCommandActions()` API, which uses `RemoteActionCompat` to handle commands using any type of component (in this case, using `PhotosPruneReceiver` as a `BroadcastReceiver` to do the small amount of work needed).